Assignment 1/requirements.txt
absl-py==1.0.0
astor==0.8.1
astunparse==1.6.3
attrdict==2.0.1
attrs==21.4.0
Babel==2.10.3
bce-python-sdk==0.8.73
beautifulsoup4==4.11.2
boto3==1.24.72
botocore==1.27.72
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
comtypes==1.1.11
contourpy==1.0.6
cssselect==1.1.0
cssutils==2.4.2
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
disjoint-set==0.7.3
distlib==0.3.4
dm-tree==0.1.7
dp-accounting==0.3.0
et-xmlfile==1.1.0
etils==0.9.0
facebook-wda==1.4.6
falcon==3.1.0
filelock==3.7.1
fire==0.5.0
flake8==4.0.1
Flask==2.1.2
Flask-Babel==2.0.0
flatbuffers==22.12.6
fonttools==4.33.3
fsspec==2023.10.0
future==0.18.2
gast==0.4.0
gevent==22.8.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
greenlet==1.1.3.post0
grpcio==1.51.1
h5py==3.7.0
identify==2.5.1
idna==3.3
imageio==2.19.3
imgaug==0.4.0
immutabledict==2.2.3
importlib-metadata==4.12.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.3
libclang==14.0.6
lmdb==1.3.0
lxml==4.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matrix==3.0.0
mccabe==0.6.1
MouseInfo==0.1.3
mpmath==1.2.1
mss==6.1.0
networkx==2.8.4
nodeenv==1.7.0
nose==1.3.7
numpy==1.19.3
oauthlib==3.2.2
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==1.4.3
parameterized==0.8.1
patsy==0.5.3
pdf2docx==0.5.6
Pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
premailer==3.10.0
private-pgm==0.0.1
promise==2.3
protobuf==3.19.1
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyclipper==1.3.0.post3
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.20.2
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
pytesseract==0.3.9
python-dateutil==2.8.2
python-docx==0.8.11
python-Levenshtein==0.12.2
pytweening==1.0.4
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
pywinauto==0.6.3
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.0
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.8.1
Shapely==1.8.2
shellcheck-py==0.8.0.4
six==1.16.0
soupsieve==2.4
statsmodels==0.13.5
sympy==1.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-metadata==1.12.0
tensorflow-probability==0.15.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.64.0
typing_extensions==4.4.0
urllib3==1.26.9
virtualenv==20.15.0
visualdl==2.2.3
Werkzeug==2.1.2
wrapt==1.14.1
XlsxWriter==3.0.8
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.5.0
